mv app/xicon/* app/* (remove xicon from url path params)
