+++ apps/f3-glossary/app/[id]/loading.tsx
