+++ apps/f3-glossary/app/region/[id]/loading.tsx
