+++ apps/f3-glossary/components/entry-layout.tsx
@@ -74,7 +74,7 @@ export function EntryLayout({ entry, related, next, prev }: EntryLayoutProps) {
 
           <div className="mt-12 flex justify-between">
             {prev ? (
-              <Link href={`/xicon/${prev.id}`}>
+              <Link href={`/${prev.id}`}>
                 <Button variant="outline" className="flex items-center gap-2">
                   <ChevronLeft className="h-4 w-4" />
                   <span>Previous</span>
@@ -85,7 +85,7 @@ export function EntryLayout({ entry, related, next, prev }: EntryLayoutProps) {
             )}
 
             {next ? (
-              <Link href={`/xicon/${next.id}`}>
+              <Link href={`/${next.id}`}>
                 <Button variant="outline" className="flex items-center gap-2">
                   <span>Next</span>
                   <ChevronRight className="h-4 w-4" />
