+++ apps/f3-glossary/components/region-detail.tsx
@@ -70,7 +70,7 @@ export function RegionDetail({ entry, related, next, prev }: RegionDetailProps) 
 
           <div className="mt-12 flex justify-between">
             {prev ? (
-              <Link href={`/xicon/${prev.id}`}>
+              <Link href={`/${prev.id}`}>
                 <Button variant="outline" className="flex items-center gap-2">
                   <ChevronLeft className="h-4 w-4" />
                   <span>Previous</span>
@@ -81,7 +81,7 @@ export function RegionDetail({ entry, related, next, prev }: RegionDetailProps) 
             )}
 
             {next ? (
-              <Link href={`/xicon/${next.id}`}>
+              <Link href={`/${next.id}`}>
                 <Button variant="outline" className="flex items-center gap-2">
                   <span>Next</span>
                   <ChevronRight className="h-4 w-4" />
