+++ apps/f3-glossary/components/term-detail.tsx
@@ -58,7 +58,7 @@ export function TermDetail({ entry, related, next, prev }: TermDetailProps) {
 
           <div className="mt-12 flex justify-between">
             {prev ? (
-              <Link href={`/xicon/${prev.id}`}>
+              <Link href={`/${prev.id}`}>
                 <Button variant="outline" className="flex items-center gap-2">
                   <ChevronLeft className="h-4 w-4" />
                   <span>Previous</span>
@@ -69,7 +69,7 @@ export function TermDetail({ entry, related, next, prev }: TermDetailProps) {
             )}
 
             {next ? (
-              <Link href={`/xicon/${next.id}`}>
+              <Link href={`/${next.id}`}>
                 <Button variant="outline" className="flex items-center gap-2">
                   <span>Next</span>
                   <ChevronRight className="h-4 w-4" />
